requirements.txt
python-dotenv
deepgram-sdk
openai
##
pyyaml
rospkg
catkin_pkg